+++ winim/inc/winimbase.nim
@@ -8,13 +8,11 @@
 import macros
 
 when not defined(noRes):
-  import os
-
   when defined(cpu64):
-    {.link: currentSourcePath().parentDir & "/../lib/winim64.res".}
+    {.link: "../lib/winim64.res".}
 
   else:
-    {.link: currentSourcePath().parentDir & "/../lib/winim32.res".}
+    {.link: "../lib/winim32.res".}
 
 macro winapi*(x: untyped): untyped =
   when not defined(noDiscardableApi):